OpenAIへのリクエスト部分を別tsファイルに切り出し #5
